refactor: centralized policy endpoints and models
